Day16 - Part 2 refactor
